+++ core/mocks/Storage.go
@@ -98,13 +98,13 @@ func (_m *Storage) CreateGlobalConfig(context storage.TransactionContext, global
 	return r0
 }
 
-// DeleteAccountPermissions provides a mock function with given fields: context, accountID, hasPermissions, permissions
-func (_m *Storage) DeleteAccountPermissions(context storage.TransactionContext, accountID string, hasPermissions bool, permissions []model.Permission) error {
-	ret := _m.Called(context, accountID, hasPermissions, permissions)
+// DeleteAccountPermissions provides a mock function with given fields: context, accountID, permissionNames
+func (_m *Storage) DeleteAccountPermissions(context storage.TransactionContext, accountID string, permissionNames []string) error {
+	ret := _m.Called(context, accountID, permissionNames)
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(storage.TransactionContext, string, bool, []model.Permission) error); ok {
-		r0 = rf(context, accountID, hasPermissions, permissions)
+	if rf, ok := ret.Get(0).(func(storage.TransactionContext, string, []string) error); ok {
+		r0 = rf(context, accountID, permissionNames)
 	} else {
 		r0 = ret.Error(0)
 	}
@@ -112,13 +112,13 @@ func (_m *Storage) DeleteAccountPermissions(context storage.TransactionContext, 
 	return r0
 }
 
-// DeleteAccountRoles provides a mock function with given fields: context, accountID, hasPermissions, roleIDs
-func (_m *Storage) DeleteAccountRoles(context storage.TransactionContext, accountID string, hasPermissions bool, roleIDs []string) error {
-	ret := _m.Called(context, accountID, hasPermissions, roleIDs)
+// DeleteAccountRoles provides a mock function with given fields: context, accountID, roleIDs
+func (_m *Storage) DeleteAccountRoles(context storage.TransactionContext, accountID string, roleIDs []string) error {
+	ret := _m.Called(context, accountID, roleIDs)
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(storage.TransactionContext, string, bool, []string) error); ok {
-		r0 = rf(context, accountID, hasPermissions, roleIDs)
+	if rf, ok := ret.Get(0).(func(storage.TransactionContext, string, []string) error); ok {
+		r0 = rf(context, accountID, roleIDs)
 	} else {
 		r0 = ret.Error(0)
 	}
@@ -261,13 +261,13 @@ func (_m *Storage) FindAccountByID(context storage.TransactionContext, id string
 	return r0, r1
 }
 
-// FindAccounts provides a mock function with given fields: limit, offset, appID, orgID, accountID, firstName, lastName, authType, authTypeIdentifier, hasPermissions, permissions, roleIDs, groupIDs
-func (_m *Storage) FindAccounts(limit int, offset int, appID string, orgID string, accountID *string, firstName *string, lastName *string, authType *string, authTypeIdentifier *string, hasPermissions *bool, permissions []string, roleIDs []string, groupIDs []string) ([]model.Account, error) {
-	ret := _m.Called(limit, offset, appID, orgID, accountID, firstName, lastName, authType, authTypeIdentifier, hasPermissions, permissions, roleIDs, groupIDs)
+// FindAccounts provides a mock function with given fields: limit, offset, appID, orgID, accountID, firstName, lastName, authType, authTypeIdentifier, anonymous, hasPermissions, permissions, roleIDs, groupIDs
+func (_m *Storage) FindAccounts(limit int, offset int, appID string, orgID string, accountID *string, firstName *string, lastName *string, authType *string, authTypeIdentifier *string, anonymous *bool, hasPermissions *bool, permissions []string, roleIDs []string, groupIDs []string) ([]model.Account, error) {
+	ret := _m.Called(limit, offset, appID, orgID, accountID, firstName, lastName, authType, authTypeIdentifier, anonymous, hasPermissions, permissions, roleIDs, groupIDs)
 
 	var r0 []model.Account
-	if rf, ok := ret.Get(0).(func(int, int, string, string, *string, *string, *string, *string, *string, *bool, []string, []string, []string) []model.Account); ok {
-		r0 = rf(limit, offset, appID, orgID, accountID, firstName, lastName, authType, authTypeIdentifier, hasPermissions, permissions, roleIDs, groupIDs)
+	if rf, ok := ret.Get(0).(func(int, int, string, string, *string, *string, *string, *string, *string, *bool, *bool, []string, []string, []string) []model.Account); ok {
+		r0 = rf(limit, offset, appID, orgID, accountID, firstName, lastName, authType, authTypeIdentifier, anonymous, hasPermissions, permissions, roleIDs, groupIDs)
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).([]model.Account)
@@ -275,8 +275,8 @@ func (_m *Storage) FindAccounts(limit int, offset int, appID string, orgID strin
 	}
 
 	var r1 error
-	if rf, ok := ret.Get(1).(func(int, int, string, string, *string, *string, *string, *string, *string, *bool, []string, []string, []string) error); ok {
-		r1 = rf(limit, offset, appID, orgID, accountID, firstName, lastName, authType, authTypeIdentifier, hasPermissions, permissions, roleIDs, groupIDs)
+	if rf, ok := ret.Get(1).(func(int, int, string, string, *string, *string, *string, *string, *string, *bool, *bool, []string, []string, []string) error); ok {
+		r1 = rf(limit, offset, appID, orgID, accountID, firstName, lastName, authType, authTypeIdentifier, anonymous, hasPermissions, permissions, roleIDs, groupIDs)
 	} else {
 		r1 = ret.Error(1)
 	}
@@ -300,6 +300,29 @@ func (_m *Storage) FindAccountsByAccountID(appID string, orgID string, accountID
 	var r1 error
 	if rf, ok := ret.Get(1).(func(string, string, []string) error); ok {
 		r1 = rf(appID, orgID, accountIDs)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
+// FindAccountsByUsername provides a mock function with given fields: context, appOrg, username
+func (_m *Storage) FindAccountsByUsername(context storage.TransactionContext, appOrg *model.ApplicationOrganization, username string) ([]model.Account, error) {
+	ret := _m.Called(context, appOrg, username)
+
+	var r0 []model.Account
+	if rf, ok := ret.Get(0).(func(storage.TransactionContext, *model.ApplicationOrganization, string) []model.Account); ok {
+		r0 = rf(context, appOrg, username)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).([]model.Account)
+		}
+	}
+
+	var r1 error
+	if rf, ok := ret.Get(1).(func(storage.TransactionContext, *model.ApplicationOrganization, string) error); ok {
+		r1 = rf(context, appOrg, username)
 	} else {
 		r1 = ret.Error(1)
 	}
@@ -1105,13 +1128,13 @@ func (_m *Storage) RegisterStorageListener(storageListener storage.Listener) {
 	_m.Called(storageListener)
 }
 
-// RemoveAccountsGroup provides a mock function with given fields: groupID, accounts, hasPermissions
-func (_m *Storage) RemoveAccountsGroup(groupID string, accounts []model.Account, hasPermissions []bool) error {
-	ret := _m.Called(groupID, accounts, hasPermissions)
+// RemoveAccountsGroup provides a mock function with given fields: groupID, accountIDs
+func (_m *Storage) RemoveAccountsGroup(groupID string, accountIDs []string) error {
+	ret := _m.Called(groupID, accountIDs)
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(string, []model.Account, []bool) error); ok {
-		r0 = rf(groupID, accounts, hasPermissions)
+	if rf, ok := ret.Get(0).(func(string, []string) error); ok {
+		r0 = rf(groupID, accountIDs)
 	} else {
 		r0 = ret.Error(0)
 	}
@@ -1133,13 +1156,13 @@ func (_m *Storage) SaveDevice(context storage.TransactionContext, device *model.
 	return r0
 }
 
-// UpdateAccountPreferences provides a mock function with given fields: accountID, preferences
-func (_m *Storage) UpdateAccountPreferences(accountID string, preferences map[string]interface{}) error {
-	ret := _m.Called(accountID, preferences)
+// UpdateAccountPreferences provides a mock function with given fields: context, accountID, preferences
+func (_m *Storage) UpdateAccountPreferences(context storage.TransactionContext, accountID string, preferences map[string]interface{}) error {
+	ret := _m.Called(context, accountID, preferences)
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(string, map[string]interface{}) error); ok {
-		r0 = rf(accountID, preferences)
+	if rf, ok := ret.Get(0).(func(storage.TransactionContext, string, map[string]interface{}) error); ok {
+		r0 = rf(context, accountID, preferences)
 	} else {
 		r0 = ret.Error(0)
 	}
@@ -1154,6 +1177,20 @@ func (_m *Storage) UpdateAccountSystemConfigs(context storage.TransactionContext
 	var r0 error
 	if rf, ok := ret.Get(0).(func(storage.TransactionContext, string, map[string]interface{}) error); ok {
 		r0 = rf(context, accountID, configs)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
+// UpdateAccountUsername provides a mock function with given fields: context, accountID, username
+func (_m *Storage) UpdateAccountUsername(context storage.TransactionContext, accountID string, username string) error {
+	ret := _m.Called(context, accountID, username)
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(storage.TransactionContext, string, string) error); ok {
+		r0 = rf(context, accountID, username)
 	} else {
 		r0 = ret.Error(0)
 	}